requirements.txt
numpy
matplotlib
opencv-python
numpy-quaternion
tqdm
joblib
jupyter
einops
omegaconf
imageio==2.9
wandb
sklearn
